refactoring composit layout logic, access control,
