+++ DemoApp/Flows/HomeFlow/AssetDetail/AssetDetailVCFactory.swift
@@ -1,0 +1,8 @@
+//
+//  AssetDetailVCFactory.swift
+//  DemoApp
+//
+//  Created by Alex Kovalov on 11/3/24.
+//
+
+import Foundation
