+++ DemoApp/Flows/HomeFlow/Home/ChilViews/PagingSectionFooterView.swift
@@ -13,7 +13,7 @@ struct PagingInfo: Equatable, Hashable {
     let currentPage: Int
 }
 
-class PagingSectionFooterView: UICollectionReusableView, ReusableView {
+final class PagingSectionFooterView: UICollectionReusableView, ReusableView {
 
     private var pagingInfoToken: AnyCancellable?
